+++ NEWS.md
@@ -1,6 +1,7 @@
 # elastes 0.1.5.9000
 
 * Changed T/F to TRUE/FALSE in function's default parameters.
+* Documented missing return values in exported functions.
 * Removed 'par()' call in 'elastes.Rmd' vignette.
 
 # elastes 0.1.5