+++ NEWS.md
@@ -1,4 +1,4 @@
-# elastes 0.1.5.9000
+# elastes 0.1.6
 
 * Changed T/F to TRUE/FALSE in function's default parameters.
 * Documented missing return values in exported functions.